+++ http4k-contract/src/test/kotlin/org/http4k/filters/contractExt.kt
@@ -1,6 +1,0 @@
-package org.http4k.filters
-
-import org.http4k.contract.ContractRenderer
-import org.http4k.filter.ServerFilters
-
-fun ServerFilters.CatchLensFailure(renderer: ContractRenderer) = CatchLensFailure { renderer.badRequest(it.failures) }