+++ http4k-format-jackson-xml/src/test/kotlin/org/http4k/filter/JacksonGenerateXmlDataClassesTest.kt
@@ -2,7 +2,7 @@ package org.http4k.filter
 
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
-import org.http4k.core.Method
+import org.http4k.core.Method.GET
 import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
 
-class GenerateXmlDataClassesTest {
+class JacksonGenerateXmlDataClassesTest {
 
     private val input = """<?xml version="1.0" encoding="UTF-8" standalone="no"?><Xml>asd<SubWithText attr="attrValue">subText</SubWithText><SubWithText attr="attrValue3">subText4</SubWithText><subWithAttr attr="attr2"/></Xml>"""
 
@@ -20,22 +20,17 @@ class GenerateXmlDataClassesTest {
 
         val os = ByteArrayOutputStream()
 
-        val body = Response(Status.OK).body(input)
-        val handler = GenerateXmlDataClasses(PrintStream(os), { 1 }).then { body }
+        val app = JacksonGenerateXmlDataClasses(PrintStream(os), { 1 }).then { Response(Status.OK).body(input) }
 
-        handler(Request(Method.GET, "/bob"))
-        val actual = String(os.toByteArray())
+        app(Request(GET, "/bob"))
 
-        assertThat(actual, equalTo("""// result generated from /bob
+        assertThat(String(os.toByteArray()), equalTo("""// result generated from /bob
 
-data class Base(val Xml: Xml?)
+data class Base(val : String?, val SubWithText: SubWithText?, val subWithAttr: SubWithAttr?)
 
 data class SubWithAttr(val attr: String?)
 
-data class SubWithText1(val attr: String?, val content: String?)
-
-data class Xml(val SubWithText: List<SubWithText1>?, val subWithAttr: SubWithAttr?, val content: String?)
+data class SubWithText(val attr: String?, val : String?)
 """))
-
     }
 }