+++ http4k-format-xml/src/test/kotlin/org/http4k/filter/GsonGenerateXmlDataClassesTest.kt
@@ -1,4 +1,4 @@
-package org.http4k.format
+package org.http4k.filter
 
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
@@ -7,12 +7,11 @@ import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status
 import org.http4k.core.then
-import org.http4k.filter.GenerateXmlDataClasses
 import org.junit.jupiter.api.Test
 import java.io.ByteArrayOutputStream
 import java.io.PrintStream
 
-class JacksonGenerateXmlDataClassesTest {
+class GsonGenerateXmlDataClassesTest {
 
     private val input = """<?xml version="1.0" encoding="UTF-8" standalone="no"?><Xml>asd<SubWithText attr="attrValue">subText</SubWithText><SubWithText attr="attrValue3">subText4</SubWithText><subWithAttr attr="attr2"/></Xml>"""
 
@@ -21,17 +20,22 @@ class JacksonGenerateXmlDataClassesTest {
 
         val os = ByteArrayOutputStream()
 
-        val app = GenerateXmlDataClasses(PrintStream(os), { 1 }).then { Response(Status.OK).body(input) }
+        val body = Response(Status.OK).body(input)
+        val handler = GsonGenerateXmlDataClasses(PrintStream(os), { 1 }).then { body }
 
-        app(Request(Method.GET, "/bob"))
+        handler(Request(Method.GET, "/bob"))
+        val actual = String(os.toByteArray())
 
-        assertThat(String(os.toByteArray()), equalTo("""// result generated from /bob
+        assertThat(actual, equalTo("""// result generated from /bob
 
-data class Base(val : String?, val SubWithText: SubWithText?, val subWithAttr: SubWithAttr?)
+data class Base(val Xml: Xml?)
 
 data class SubWithAttr(val attr: String?)
 
-data class SubWithText(val attr: String?, val : String?)
+data class SubWithText1(val attr: String?, val content: String?)
+
+data class Xml(val SubWithText: List<SubWithText1>?, val subWithAttr: SubWithAttr?, val content: String?)
 """))
+
     }
 }